chore: move real proof to example
