+++ poseidon/src/circuit.rs
@@ -564,25 +564,4 @@ mod tests {
         let prover = MockProver::run(k, &circuit, vec![]).expect("cannot prove");
         assert_eq!(prover.verify(), Ok(()))
     }
-
-    #[test]
-    fn test_poseidon_circuit_real_prover() {
-        let rng = OsRng;
-
-        let message = [Fr::random(rng), Fr::random(rng)];
-        let output = Hash::<Fr, OrchardNullifier, ConstantLength<2>, 3, 2>::init().hash(message);
-
-        let k = 6;
-        let circuit = PoseidonCircuit::<OrchardNullifier, Fr, ConstantLength<2>, 3, 2, 2> {
-            message,
-            output,
-            _marker: PhantomData,
-            _marker2: PhantomData,
-        };
-
-        // Test with prover
-        let mut prover = PoseidonProver::new(k, circuit, true);
-        let proof = prover.create_proof();
-        assert!(prover.verify(proof));
-    }
 }
