+++ zkmemory/src/constraints/sorted_memory_circuit.rs
@@ -434,7 +434,7 @@ mod test {
     extern crate alloc;
     extern crate std;
     use alloc::{vec, vec::Vec};
-    use std::marker::PhantomData;
+    use core::marker::PhantomData;
     // Common test function to build and the the SortedMemoryCircuit
     fn build_and_test_circuit(trace: Vec<ConvertedTraceRecord<Fp>>, k: u32) {
         let circuit = SortedMemoryCircuit::<Fp> {
